+++ src/crypto/avs_crypto_utils.c
@@ -701,6 +701,16 @@ const avs_crypto_pki_x509_name_key_t AVS_CRYPTO_PKI_X509_NAME_CN = {
     .value_id_octet = 0x0C // UTF8String
 };
 
+const avs_crypto_pki_x509_name_key_t AVS_CRYPTO_PKI_X509_NAME_O = {
+    .oid = (const avs_crypto_asn1_oid_t *) "\x06\x03\x55\x04\x0a",
+    .value_id_octet = 0x0C // UTF8String
+};
+
+const avs_crypto_pki_x509_name_key_t AVS_CRYPTO_PKI_X509_NAME_C = {
+    .oid = (const avs_crypto_asn1_oid_t *) "\x06\x03\x55\x04\x06",
+    .value_id_octet = 0x0C // UTF8String
+};
+
 bool _avs_crypto_aead_parameters_valid(size_t key_len,
                                        size_t iv_len,
                                        size_t tag_len) {